Reading mockintosh.yml file
