+++ tests/docker-compose.yml
@@ -26,6 +26,24 @@ services:
       - /var/run:rw,noexec,nosuid
     hostname: edge-router
     restart: always
+  edge-routerhttp:
+    image: weaveworksdemos/edge-router:0.1.1
+    ports:
+      - '80:80'
+      - '8080:8080'
+    cap_drop:
+      - all
+    cap_add:
+      - NET_BIND_SERVICE
+      - CHOWN
+      - SETGID
+      - SETUID
+      - DAC_OVERRIDE
+    read_only: true
+    tmpfs:
+      - /var/run:rw,noexec,nosuid
+    hostname: edge-routerhttp
+    restart: always
   catalogue:
     image: weaveworksdemos/catalogue:0.3.5
     hostname: catalogue
