--- ventilator/configurator/cli.py
@@ -1,5 +1,0 @@
-def CLI():
-    """
-        Using CLI
-    """
-    pass
